requirements.txt
alabaster==0.7.10
alembic==1.0.6
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astroid==1.6.1
astropy==2.0.3
atomicwrites==1.2.1
attrs==17.4.0
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.6
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.2
blinker==1.4
bokeh==0.12.13
boto==2.48.0
boto3==1.12.26
botocore==1.15.26
Bottleneck==1.2.1
cachetools==3.1.1
certifi==2019.11.28
cffi==1.11.4
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
conda==4.4.10
conda-build==3.4.1
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.1.4
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
decorator==4.2.1
distributed==1.20.2
docutils==0.14
entrypoints==0.2.3
environs==4.0.0
et-xmlfile==1.0.1
factory-boy==2.11.1
Faker==1.0.2
fastcache==1.0.2
filelock==2.0.13
flake8==3.5.0
flake8-blind-except==0.1.1
flake8-debugger==3.1.0
flake8-docstrings==1.3.0
flake8-isort==2.5
flake8-polyfill==1.0.2
flake8-quotes==1.0.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.4.0
Flask-Cors==3.0.3
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-Webpack==0.1.0
Flask-WTF==0.14.2
gevent==1.2.2
glob2==0.6
gmpy2==2.0.8
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
greenlet==0.4.12
grpcio==1.24.0
gunicorn==19.9.0
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
imageio==2.2.0
imagesize==0.7.1
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.5
jupyterlab-launcher==0.10.2
keyboard==0.13.2
lazy-object-proxy==1.3.1
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.18.0
matplotlib==2.1.2
mccabe==0.6.1
mistune==0.8.3
more-itertools==5.0.0
mpmath==1.0.0
msgpack-python==0.5.1
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
numba==0.36.2
numexpr==2.6.4
numpy==1.14.0
numpydoc==0.7.0
oauthlib==3.1.0
odo==0.5.1
olefile==0.45.1
opencv-python==4.1.2.30
openpyxl==2.4.10
packaging==16.8
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.1
pep8-naming==0.7.0
pexpect==4.3.1
pickleshare==0.7.4
Pillow==6.2.1
pkginfo==1.4.1
pluggy==0.8.1
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.9.2
psutil==5.4.3
psycopg2==2.7.5
ptyprocess==0.5.2
py==1.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyAutoGUI==0.9.38
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pydocstyle==3.0.0
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.2
PyMsgBox==1.0.6
pynput==1.4
pyobjc==5.1.1
pyobjc-core==5.1.1
pyobjc-framework-Accounts==5.1.1
pyobjc-framework-AddressBook==5.1.1
pyobjc-framework-AppleScriptKit==5.1.1
pyobjc-framework-AppleScriptObjC==5.1.1
pyobjc-framework-ApplicationServices==5.1.1
pyobjc-framework-Automator==5.1.1
pyobjc-framework-AVFoundation==5.1.1
pyobjc-framework-AVKit==5.1.1
pyobjc-framework-CalendarStore==5.1.1
pyobjc-framework-CFNetwork==5.1.1
pyobjc-framework-CloudKit==5.1.1
pyobjc-framework-Cocoa==5.1.1
pyobjc-framework-Collaboration==5.1.1
pyobjc-framework-ColorSync==5.1.1
pyobjc-framework-Contacts==5.1.1
pyobjc-framework-ContactsUI==5.1.1
pyobjc-framework-CoreAudio==5.1.1
pyobjc-framework-CoreAudioKit==5.1.1
pyobjc-framework-CoreBluetooth==5.1.1
pyobjc-framework-CoreData==5.1.1
pyobjc-framework-CoreLocation==5.1.1
pyobjc-framework-CoreMedia==5.1.1
pyobjc-framework-CoreMediaIO==5.1.1
pyobjc-framework-CoreML==5.1.1
pyobjc-framework-CoreServices==5.1.1
pyobjc-framework-CoreSpotlight==5.1.1
pyobjc-framework-CoreText==5.1.1
pyobjc-framework-CoreWLAN==5.1.1
pyobjc-framework-CryptoTokenKit==5.1.1
pyobjc-framework-DictionaryServices==5.1.1
pyobjc-framework-DiscRecording==5.1.1
pyobjc-framework-DiscRecordingUI==5.1.1
pyobjc-framework-DiskArbitration==5.1.1
pyobjc-framework-DVDPlayback==5.1.1
pyobjc-framework-EventKit==5.1.1
pyobjc-framework-ExceptionHandling==5.1.1
pyobjc-framework-ExternalAccessory==5.1.1
pyobjc-framework-FinderSync==5.1.1
pyobjc-framework-FSEvents==5.1.1
pyobjc-framework-GameCenter==5.1.1
pyobjc-framework-GameController==5.1.1
pyobjc-framework-GameKit==5.1.1
pyobjc-framework-GameplayKit==5.1.1
pyobjc-framework-ImageCaptureCore==5.1.1
pyobjc-framework-IMServicePlugIn==5.1.1
pyobjc-framework-InputMethodKit==5.1.1
pyobjc-framework-InstallerPlugins==5.1.1
pyobjc-framework-InstantMessage==5.1.1
pyobjc-framework-Intents==5.1.1
pyobjc-framework-IOSurface==5.1.1
pyobjc-framework-iTunesLibrary==5.1.1
pyobjc-framework-LatentSemanticMapping==5.1.1
pyobjc-framework-LaunchServices==5.1.1
pyobjc-framework-libdispatch==5.1.1
pyobjc-framework-LocalAuthentication==5.1.1
pyobjc-framework-MapKit==5.1.1
pyobjc-framework-MediaAccessibility==5.1.1
pyobjc-framework-MediaLibrary==5.1.1
pyobjc-framework-MediaPlayer==5.1.1
pyobjc-framework-MediaToolbox==5.1.1
pyobjc-framework-ModelIO==5.1.1
pyobjc-framework-MultipeerConnectivity==5.1.1
pyobjc-framework-NetFS==5.1.1
pyobjc-framework-NetworkExtension==5.1.1
pyobjc-framework-NotificationCenter==5.1.1
pyobjc-framework-OpenDirectory==5.1.1
pyobjc-framework-OSAKit==5.1.1
pyobjc-framework-Photos==5.1.1
pyobjc-framework-PhotosUI==5.1.1
pyobjc-framework-PreferencePanes==5.1.1
pyobjc-framework-PubSub==5.1.1
pyobjc-framework-QTKit==5.1.1
pyobjc-framework-Quartz==5.1.1
pyobjc-framework-SafariServices==5.1.1
pyobjc-framework-SceneKit==5.1.1
pyobjc-framework-ScreenSaver==5.1.1
pyobjc-framework-ScriptingBridge==5.1.1
pyobjc-framework-SearchKit==5.1.1
pyobjc-framework-Security==5.1.1
pyobjc-framework-SecurityFoundation==5.1.1
pyobjc-framework-SecurityInterface==5.1.1
pyobjc-framework-ServiceManagement==5.1.1
pyobjc-framework-Social==5.1.1
pyobjc-framework-SpriteKit==5.1.1
pyobjc-framework-StoreKit==5.1.1
pyobjc-framework-SyncServices==5.1.1
pyobjc-framework-SystemConfiguration==5.1.1
pyobjc-framework-VideoToolbox==5.1.1
pyobjc-framework-Vision==5.1.1
pyobjc-framework-WebKit==5.1.1
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
pypresence==2.1.1
PyScreeze==0.1.18
PySocks==1.6.7
pystray==0.14.3
pytesseract==0.3.0
pytest==3.7.4
python-dateutil==2.6.1
python-dotenv==0.10.1
python-editor==1.0.3
PyTweening==1.0.3
pytz==2019.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.22.0
requests-aws-sign==0.1.5
requests-oauthlib==1.3.0
requests-sigv4==0.1.5
rope==0.10.7
rsa==4.0
ruamel-yaml==0.15.35
ruamel.yaml==0.15.7
s3transfer==0.3.3
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.4.2
sense-emu==1.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.13.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.2.11
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tabulate==0.8.6
tblib==1.3.2
terminado==0.8.1
testfixtures==6.4.3
testpath==0.3.1
text-unidecode==1.2
toolz==0.9.0
tornado==4.5.3
traitlets==4.3.2
typing==3.6.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.7
waitress==1.2.0
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.8.5
WebTest==2.0.30
Werkzeug==0.14.1
widgetsnbextension==3.1.0
wrapt==1.10.11
WTForms==2.2.1
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.2.0
zict==0.1.3
